+++ services/objectstorage/CHANGELOG.md
@@ -1,3 +1,10 @@
+## v0.2.0
+- New model classes: `ComplianceLockResponse`, `CredentialsGroupExtended`, `DefaultRetentionResponse`, `DeleteDefaultRetentionResponse`, `GetCredentialsGroupResponse`, `SetDefaultRetentionPayload`
+- New methods for `Bucket` model class: `objectLockEnabled`,`setObjectLockEnabled`,`getObjectLockEnabled`
+- New enum: `RetentionMode`
+- New API client methods: `createComplianceLock`, `deleteComplianceLock`, `deleteDefaultRetention`, `getComplianceLock`, `getCredentialsGroup`, `getDefaultRetention`, `setDefaultRetention`
+- **Breaking change:** New parameter `objectLockEnabled` for `createBucket` API client method
+
 ## v0.1.3
 - Bump dependency `org.apache.commons:commons-lang3` to `3.18.0`
 - Bump dependency `org.openapitools:jackson-databind-nullable` to `0.2.8`